Split motion service test files
